Comment testin phpunit

+++ app/controllers/admin/CommentController.php
@@ -5,8 +5,8 @@ use App\Models\Article;
 use App\Models\User;
 use Input, Notification, Redirect, Sentry, Str;
 
-class CommentController extends \BaseController {
-
+class CommentController extends \BaseController 
+{
 	/**
 	 * Display a listing of the resource.
 	 *
@@ -24,14 +24,19 @@ class CommentController extends \BaseController {
 	 * @return Response
 	 */
 	public function store()
-	{
-		$comment = Comment::create(array(
+	{	
+		$new_comment = [
 			'author' => Input::get('author'),
 			'text' => Input::get('text'),
 			'article_id' => Input::get('article_id')
-		));
+		];
 
-		return $comment;
+		if(in_array('', $new_comment))
+		{
+			throw new \PDOException('400, Invalid Data Format for Comment');
+		}
+
+		return Comment::create($new_comment);;
 	}
 
 	// Update comment.
@@ -53,15 +58,24 @@ class CommentController extends \BaseController {
 	 */
 	public function destroy($id)
 	{
-		$article = Comment::find($id)->article;
+		if(!$comment = Comment::find($id))
+		{			
+			throw new \Symfony\Component\HttpKernel\Exception\NotFoundHttpException("Delete was failed. Comment doesn't exist.");
+		}
+
 		Comment::destroy($id);
-		return $article;
+		return $comment->article;
 	}
 
 
 	public function show($slug)
-    {    	
-    	$comment = Article::where('slug', $slug)->first()->comments();
-        return $comment->get();
+    {    
+    	if (!$article = Article::where('slug', $slug)->first()) 
+		{
+			throw new \Symfony\Component\HttpKernel\Exception\NotFoundHttpException('404, Comments not found');
+		}
+
+    	$comment = $article->comments();
+        return $comment->get();	    
     }
 }
